+++ Projectile.cpp
@@ -7,7 +7,7 @@
 // todo: color based on isPlayer
 Projectile::Projectile() : Projectile(0, sf::Vector2(0.f, 0.f), sf::Color::Black, 0, sf::Vector2(0.f, 0.f), false, 0) {}
 
-Projectile::Projectile(int r, sf::Vector2f position, sf::Color Colour, int speed,
+Projectile::Projectile(int r, sf::Vector2f position, sf::Color Colour, float speed,
                        sf::Vector2f destination, bool isPlayer, int damage)
     : Entity(r, position, isPlayer ? sf::Color::Cyan : sf::Color::Red,
              speed, destination, isPlayer),
